+++ tests/Imagine/Test/Gd/DrawerTest.php
@@ -28,6 +28,51 @@ class DrawerTest extends AbstractDrawerTest
         }
     }
 
+    /**
+     * @dataProvider thicknessAndFillProvider
+     *
+     * {@inheritdoc}
+     *
+     * @see \Imagine\Test\Draw\AbstractDrawerTest::testChord()
+     */
+    public function testChord($thickness, $fill)
+    {
+        if ($fill) {
+            $this->markTestSkipped('The GD Drawer can NOT draw correctly filled chords');
+        }
+        parent::testChord($thickness, $fill);
+    }
+
+    /**
+     * @dataProvider thicknessAndFillProvider
+     *
+     * {@inheritdoc}
+     *
+     * @see \Imagine\Test\Draw\AbstractDrawerTest::testCircle()
+     */
+    public function testCircle($thickness, $fill)
+    {
+        if (!$fill && $thickness > 1) {
+            $this->markTestSkipped('The GD Drawer can NOT draw correctly not filled circles with a thickness greater than 1');
+        }
+        parent::testCircle($thickness, $fill);
+    }
+
+    /**
+     * @dataProvider thicknessAndFillProvider
+     *
+     * {@inheritdoc}
+     *
+     * @see \Imagine\Test\Draw\AbstractDrawerTest::testEllipse()
+     */
+    public function testEllipse($thickness, $fill)
+    {
+        if (!$fill && $thickness > 1) {
+            $this->markTestSkipped('The GD Drawer can NOT draw correctly not filled ellipses with a thickness greater than 1');
+        }
+        parent::testEllipse($thickness, $fill);
+    }
+
     protected function getImagine()
     {
         return new Imagine();