+++ CHANGELOG.md
@@ -3,7 +3,10 @@
 ### NEXT (YYYY-MM-DD)
 * New `FontInterface` method: `wrapText` - split a text into multiple lines, so that it fits a specific width (@mlocati)  
   **BREAKING CHANGE** if you have your own `FontInterface` implementation, it now must implement `wrapText`
-
+* Drawer methods can now accept a thickness of zero (@mlocati)
+* Fix drawing unfilled chords with GD driver (@mlocati)
+* Fix thickness drawing of unfilled chords with Imagick and Gmagick drivers (@mlocati)
+* Fix handling of radius in `circle` method implementations (@mlocati)
 
 ### 1.0.0-alpha2 (2018-09-08)
 * The `coalesce` method of `LayerInterface` instances now returns the LayerInterface itself (@mlocati)